Meetings with Puch AI/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
distro==1.9.0
exceptiongroup==1.3.0
fastmcp==2.8.1
future==1.0.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
mcp==1.9.4
mdurl==0.1.2
oauthlib==3.3.0
openai==1.86.0
openapi-pydantic==0.5.1
pdfminer.six==20250506
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-multipart==0.0.20
readabilipy==0.3.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.47.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
webencodings==0.5.1
